Refactor into using threading.Thread and queue.Queue (again)
The multiprocessing approach would not enable us to interrupt
the normal ledstrip operation. By running the ModLedController
as a threading.Thread and using a queue.Queue for communicating
from the Twisted MainThread, we get the flexibility to react
to new values in the queue, on which we can operate.

We've chosen to keep the data that is interchange simple: it more or
less functions only as a pass through for the control_signal
that is triggered when a value is written to the Modbus server
context. The trigger is forwarded to the ModLedController via the
queue, which will result in an exception being raised in the show()
function of the ledstrip, overridden from the Adafruit_Ledstrip
implementation, resulting in control being handed back to the
ModLedController.

The logic for reinitializing the ledstrip needs to be written, but
should be fairly similar to the logic in the ledstrip.py example.